Add docstrings to conftests
